chore: Re-organized components folder
