--- ui/src/components/ContactList/ContactList.module.css
@@ -1,20 +1,0 @@
-.container {
-  padding: 1rem 0;
-}
-
-.body {
-  display: flex;
-  gap: 1rem;
-  justify-content: space-between;
-  margin: 1rem 0;
-}
-
-.item {
-  outline: none;
-  border: none;
-  @apply p-2 rounded bg-muted text-muted-foreground;
-}
-
-.selected {
-  @apply bg-foreground text-background;
-}
